turn list page filter stuff into a hook
