+++ txcdr-client/utils/hooks/useFilterController.tsx
@@ -1,0 +1,104 @@
+import { PostgrestSingleResponse } from "@supabase/supabase-js";
+import { useGlobalSearchParams } from "expo-router";
+import { useState, useEffect } from "react";
+import { Alert } from "react-native";
+import {
+  AddressQueryParams,
+  Filter,
+  FilterController,
+} from "../../components/FilterController";
+import { Tables } from "../../types/supabase";
+import { supabase } from "../supabase";
+import { useRole } from "./useRole";
+import { useUser } from "./useUser";
+
+export function useFilterController(filters: Filter[]) {
+  const params = useGlobalSearchParams<Partial<AddressQueryParams>>();
+  const [addresses, setAddresses] = useState<Tables<"EventAddress">[]>();
+  const [zipCodes, setZipCodes] = useState<string[]>([]);
+  const [blockIds, setBlockIds] = useState<string[]>([]);
+  const [eventIds, setEventIds] = useState<number[]>([]);
+  const user = useUser();
+  const [role] = useRole();
+  const [registeredEventIds, setRegisteredEventIds] = useState<number[]>();
+
+  // Load events the user is registered for
+  useEffect(() => {
+    const func = async () => {
+      if (user && role != "USER") {
+        const res = await supabase
+          .from("EventVolunteer")
+          .select("eventId")
+          .eq("volunteerId", user.id)
+          .eq("approved", true);
+        if (res.error) {
+          console.log(res.error);
+          Alert.alert("Failed to fetch events", res.error.message);
+          return;
+        }
+        setRegisteredEventIds(res.data.map((e) => e.eventId));
+      }
+    };
+
+    func();
+  }, [user, role]);
+
+  // Load addresses
+  useEffect(() => {
+    const func = async () => {
+      let res: PostgrestSingleResponse<Tables<"EventAddress">[]> | undefined;
+      if (role != "USER") {
+        // Admins can see all addresses
+        res = await supabase.from("EventAddress").select("*");
+      } else {
+        // Users can see only the addresses of events they are registered for
+        res = await supabase
+          .from("EventAddress")
+          .select("*")
+          .in("eventId", registeredEventIds!);
+      }
+      if (res) {
+        if (res.error) {
+          console.log(res.error);
+          Alert.alert("Failed to fetch addresses", res.error.message);
+          return;
+        }
+        setAddresses(res.data);
+
+        // Build data for the filters
+        const zipCodes: Set<string> = new Set();
+        const blockIds: Set<string> = new Set();
+        const eventIds: Set<number> = new Set();
+        for (const a of res.data) {
+          zipCodes.add(a.zipCode);
+          blockIds.add(a.blockId);
+          eventIds.add(a.eventId);
+        }
+        setZipCodes(Array.from(zipCodes));
+        setBlockIds(Array.from(blockIds));
+        setEventIds(Array.from(eventIds));
+      }
+    };
+
+    if (
+      addresses == undefined &&
+      user &&
+      role &&
+      registeredEventIds != undefined
+    )
+      func();
+  }, [role, user, registeredEventIds, addresses]);
+
+  return {
+    controller: (
+      <FilterController
+        filters={filters}
+        zipCodes={zipCodes}
+        blockIds={blockIds}
+        eventIds={eventIds.map((id) => id.toString())}
+      />
+    ),
+    params: params,
+    addresses: addresses,
+  };
+}
